Move windows build from appveyor to travis

+++ script/travis/install.sh
@@ -24,29 +24,45 @@ source ./script/travis/travis_retry.sh
 
 source ./script/travis/set.sh
 
-travis_retry apt-get -y --quiet update
-travis_retry apt-get -y install sudo
+: ${ALPAKA_CI_ANALYSIS?"ALPAKA_CI_ANALYSIS must be specified"}
+: ${ALPAKA_ACC_GPU_CUDA_ENABLE?"ALPAKA_ACC_GPU_CUDA_ENABLE must be specified"}
+: ${ALPAKA_ACC_GPU_HIP_ENABLE?"ALPAKA_ACC_GPU_HIP_ENABLE must be specified"}
+: ${ALPAKA_HIP_PLATFORM?"ALPAKA_HIP_PLATFORM must be specified"}
 
-# software-properties-common: 'add-apt-repository' and certificates for wget https download
-# binutils: ld
-# xz-utils: xzcat
-travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install software-properties-common wget git make binutils xz-utils
-
-./script/travis/install_cmake.sh
-if [ "${ALPAKA_CI_ANALYSIS}" == "ON" ] ;then ./script/travis/install_analysis.sh ;fi
-# Install CUDA before installing gcc as it installs gcc-4.8 and overwrites our selected compiler
-if [ "${ALPAKA_ACC_GPU_CUDA_ENABLE}" == "ON" ] || [ "${ALPAKA_ACC_GPU_HIP_ENABLE}" == "ON" ] && [ "${ALPAKA_HIP_PLATFORM}" == "nvcc" ]
+if [ "$TRAVIS_OS_NAME" = "linux" ]
 then
-    ./script/travis/install_cuda.sh
+    travis_retry apt-get -y --quiet update
+    travis_retry apt-get -y install sudo
+
+    # software-properties-common: 'add-apt-repository' and certificates for wget https download
+    # binutils: ld
+    # xz-utils: xzcat
+    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install software-properties-common wget git make binutils xz-utils
+
+    ./script/travis/install_cmake.sh
+    if [ "${ALPAKA_CI_ANALYSIS}" == "ON" ] ;then ./script/travis/install_analysis.sh ;fi
+    # Install CUDA before installing gcc as it installs gcc-4.8 and overwrites our selected compiler
+    if [ "${ALPAKA_ACC_GPU_CUDA_ENABLE}" == "ON" ] || [ "${ALPAKA_ACC_GPU_HIP_ENABLE}" == "ON" ] && [ "${ALPAKA_HIP_PLATFORM}" == "nvcc" ]
+    then
+        ./script/travis/install_cuda.sh
+    fi
+    if [ "${CXX}" == "g++" ] ;then ./script/travis/install_gcc.sh ;fi
+    if [ "${CXX}" == "clang++" ] ;then source ./script/travis/install_clang.sh ;fi
+    if [ "${ALPAKA_ACC_GPU_HIP_ENABLE}" == "ON" ] ;then ./script/travis/install_hip.sh ;fi
 fi
-if [ "${CXX}" == "g++" ] ;then ./script/travis/install_gcc.sh ;fi
-if [ "${CXX}" == "clang++" ] ;then source ./script/travis/install_clang.sh ;fi
+
 # If the variable is not set, the backend will most probably be used by default so we install it.
-if [[ ! -v ALPAKA_ACC_CPU_B_TBB_T_SEQ_ENABLE || "${ALPAKA_ACC_CPU_B_TBB_T_SEQ_ENABLE}" == "ON" ]] ;then ./script/travis/install_tbb.sh ;fi
-if [ "${ALPAKA_ACC_GPU_HIP_ENABLE}" == "ON" ] ;then ./script/travis/install_hip.sh ;fi
+if [[ ! -v ALPAKA_ACC_CPU_B_TBB_T_SEQ_ENABLE || "${ALPAKA_ACC_CPU_B_TBB_T_SEQ_ENABLE}" == "ON" ]]
+then
+    ./script/travis/install_tbb.sh
+fi
+
 ./script/travis/install_boost.sh
 
-# Minimize docker image size
-sudo apt-get --quiet --purge autoremove
-sudo apt-get clean
-rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/*
+if [ "$TRAVIS_OS_NAME" = "linux" ]
+then
+    # Minimize docker image size
+    sudo apt-get --quiet --purge autoremove
+    sudo apt-get clean
+    rm -rf /var/lib/apt/lists/* /tmp/* /var/tmp/*
+fi
